utility/math: Use double precision for normalise().

+++ utility/math.hpp
@@ -4,11 +4,11 @@
 namespace oodles {
 
 template<class T>
-float normalise(const T &n, // Arbitrary value
-                const T &to_min, // Target range
-                const T &to_max,
-                const T &from_min, // Source range
-                const T &from_max);
+double normalise(const T &n, // Arbitrary value
+                 const T &to_min, // Target range
+                 const T &to_max,
+                 const T &from_min, // Source range
+                 const T &from_max);
 
 } // oodles
